+++ skbio/tree/tests/test_tree.py
@@ -18,6 +18,7 @@ from skbio.io._fileobject import StringIO
 from skbio import DistanceMatrix, TreeNode
 from skbio.tree import (DuplicateNodeError, NoLengthError,
                         TreeError, MissingNodeError, NoParentError)
+from skbio.util import RepresentationWarning
 
 
 class TreeTests(TestCase):
@@ -607,8 +608,22 @@ class TreeTests(TestCase):
 
     def test_tip_tip_distances_no_length(self):
         t = TreeNode.read(StringIO(u"((a,b)c,(d,e)f);"))
-        with self.assertRaises(NoLengthError):
-            t.tip_tip_distances()
+        exp_t = TreeNode.read(StringIO(u"((a:0,b:0)c:0,(d:0,e:0)f:0);"))
+        exp_t_dm = exp_t.tip_tip_distances()
+
+        t_dm = nptest.assert_warns(RepresentationWarning, t.tip_tip_distances)
+        self.assertEqual(t_dm, exp_t_dm)
+
+        for node in t.preorder():
+            self.assertIs(node.length, None)
+
+    def test_tip_tip_distances_missing_length(self):
+        t = TreeNode.read(StringIO(u"((a,b:6)c:4,(d,e:0)f);"))
+        exp_t = TreeNode.read(StringIO(u"((a:0,b:6)c:4,(d:0,e:0)f:0);"))
+        exp_t_dm = exp_t.tip_tip_distances()
+
+        t_dm = nptest.assert_warns(RepresentationWarning, t.tip_tip_distances)
+        self.assertEqual(t_dm, exp_t_dm)
 
     def test_neighbors(self):
         """Get neighbors of a node"""