+++ skbio/util/__init__.py
@@ -52,6 +52,7 @@ Warnings
    :toctree: generated/
 
    EfficiencyWarning
+   RepresentationWarning
 
 """
 
@@ -65,7 +66,7 @@ Warnings
 
 from __future__ import absolute_import, division, print_function
 
-from ._warning import EfficiencyWarning
+from ._warning import EfficiencyWarning, RepresentationWarning
 from ._exception import TestingUtilError
 from ._misc import (cardinal_to_ordinal, create_dir, find_duplicates, flatten,
                     is_casava_v180_or_later, remove_files, safe_md5)
@@ -73,8 +74,8 @@ from ._testing import (get_data_path, TestRunner,
                        assert_ordination_results_equal,
                        assert_data_frame_almost_equal)
 
-__all__ = ['EfficiencyWarning', 'TestingUtilError', 'cardinal_to_ordinal',
-           'create_dir', 'find_duplicates', 'flatten',
+__all__ = ['EfficiencyWarning', 'RepresentationWarning', 'TestingUtilError',
+           'cardinal_to_ordinal', 'create_dir', 'find_duplicates', 'flatten',
            'is_casava_v180_or_later', 'remove_files', 'safe_md5',
            'get_data_path', 'TestRunner', 'assert_ordination_results_equal',
            'assert_data_frame_almost_equal']